Refactored views into more reasonable parts (removed prior changes to unrelated files in #71)
