+++ WebCLI/urls.py
@@ -15,9 +15,11 @@ Including another URLconf
 """
 # pages/urls.py
 from django.urls import path
-from .views import new_algorithm, algorithm_details_view, new_algorithm_type
-from .views import new_molecule, SignUpView, AlgorithmListView, add_metrics
-from .views import MyAlgorithmListView, add_version, compare_algorithms, update_algorithm
+from .views.listings import MyAlgorithmListView, AlgorithmListView
+from .views.views import new_algorithm, algorithm_details_view, new_algorithm_type
+from .views.views import new_molecule, add_metrics
+from .views.views import add_version, compare_algorithms, update_algorithm
+from .forms import SignUpView
 
 urlpatterns = [
     path('', AlgorithmListView.as_view(), name='home'),
